requirements.txt
mutagen
pillow
requests